+++ compiler/include/expr.h
@@ -185,118 +185,7 @@ class UnresolvedSymExpr : public Expr {
 };
 
 
-enum TryTag {
-  TRY_TAG_NONE,
-  TRY_TAG_IN_TRY,
-  TRY_TAG_IN_TRYBANG
-};
-
-// Note -- isCallExpr() returns true for CallExpr and also
-// ContextCallExpr. Therefore, it is important to use toCallExpr()
-// instead of casting to CallExpr* directly.
-class CallExpr : public Expr {
-public:
-  PrimitiveOp* primitive;        // primitive expression (baseExpr == NULL)
-  Expr*        baseExpr;         // function expression
-
-  AList        argList;          // function actuals
-
-  bool         partialTag;
-  bool         methodTag;        // Set to true if the call is a method call.
-  bool         square;           // true if call made with square brackets
-  TryTag       tryTag;
-
-  CallExpr(BaseAST*     base,
-           BaseAST*     arg1 = NULL,
-           BaseAST*     arg2 = NULL,
-           BaseAST*     arg3 = NULL,
-           BaseAST*     arg4 = NULL,
-           BaseAST*     arg5 = NULL);
-
-  CallExpr(PrimitiveOp* prim,
-           BaseAST*     arg1 = NULL,
-           BaseAST*     arg2 = NULL,
-           BaseAST*     arg3 = NULL,
-           BaseAST*     arg4 = NULL,
-           BaseAST*     arg5 = NULL);
-
-  CallExpr(PrimitiveTag prim,
-           BaseAST*     arg1 = NULL,
-           BaseAST*     arg2 = NULL,
-           BaseAST*     arg3 = NULL,
-           BaseAST*     arg4 = NULL,
-           BaseAST*     arg5 = NULL);
-
-  CallExpr(const char*  name,
-           BaseAST*     arg1 = NULL,
-           BaseAST*     arg2 = NULL,
-           BaseAST*     arg3 = NULL,
-           BaseAST*     arg4 = NULL,
-           BaseAST*     arg5 = NULL);
-
-  ~CallExpr();
-
-  virtual void    verify();
-
-  DECLARE_COPY(CallExpr);
-
-
-  virtual void    accept(AstVisitor* visitor);
-
-  virtual void    replaceChild(Expr* old_ast, Expr* new_ast);
-
-  virtual GenRet  codegen();
-  virtual void    prettyPrint(std::ostream* o);
-  virtual QualifiedType qualType();
-
-  virtual Expr*   getFirstChild();
-
-  virtual Expr*   getFirstExpr();
-  virtual Expr*   getNextExpr(Expr* expr);
-
-  void            insertAtHead(BaseAST* ast);
-  void            insertAtTail(BaseAST* ast);
-
-  // True if the callExpr has been emptied (aka dead)
-  bool            isEmpty()                                              const;
-
-  bool            isCast();
-  Expr*           castFrom();
-  Expr*           castTo();
-
-  bool            isPrimitive()                                          const;
-  bool            isPrimitive(PrimitiveTag primitiveTag)                 const;
-  bool            isPrimitive(const char*  primitiveName)                const;
-
-  void            setUnresolvedFunction(const char* name);
-
-  bool            isResolved()                                           const;
-  FnSymbol*       resolvedFunction()                                     const;
-  void            setResolvedFunction(FnSymbol* fn);
-  FnSymbol*       resolvedOrVirtualFunction()                            const;
-
-  FnSymbol*       theFnSymbol()                                          const;
-
-  bool            isNamed(const char*)                                   const;
-  bool            isNamedAstr(const char*)                               const;
-
-  int             numActuals()                                           const;
-  Expr*           get(int index)                                         const;
-  FnSymbol*       findFnSymbol();
-
-  void            convertToNoop();
-
-private:
-  GenRet          codegenPrimitive();
-  GenRet          codegenPrimMove();
-
-  void            codegenInvokeOnFun();
-  void            codegenInvokeTaskFun(const char* name);
-
-  GenRet          codegenBasicPrimitiveExpr()                            const;
-
-  bool            isRefExternStarTuple(Symbol* formal, Expr* actual)     const;
-};
+#include "CallExpr.h"
 
 
 //
@@ -456,15 +345,6 @@ static inline bool isTaskFun(FnSymbol* fn) {
          fn->hasFlag(FLAG_ON);
 }
 
-static inline FnSymbol* resolvedToTaskFun(CallExpr* call) {
-  INT_ASSERT(call);
-  if (FnSymbol* cfn = call->resolvedFunction()) {
-    if (isTaskFun(cfn))
-      return cfn;
-  }
-  return NULL;
-}
-
 // Does this function require "capture for parallelism"?
 // Yes, if it comes from a begin/cobegin/coforall block in Chapel source.
 static inline bool needsCapture(FnSymbol* taskFn) {
@@ -492,10 +372,6 @@ bool get_uint(Expr *e, uint64_t *i); // false is failure
 bool get_string(Expr *e, const char **s); // false is failure
 const char* get_string(Expr* e); // fatal on failure
 
-CallExpr* callChplHereAlloc(Type* type, VarSymbol* md = NULL);
-void insertChplHereAlloc(Expr *call, bool insertAfter, Symbol *sym,
-                         Type* t, VarSymbol* md = NULL);
-CallExpr* callChplHereFree(BaseAST* p);
 
 // Walk the subtree of expressions rooted at "expr" in postorder, returning the
 // current expression in "e", stopping after "expr" has been returned.
@@ -508,17 +384,17 @@ CallExpr* callChplHereFree(BaseAST* p);
 
 Expr* getNextExpr(Expr* expr);
 
-CallExpr* createCast(BaseAST* src, BaseAST* toType);
-
 Expr* new_Expr(const char* format, ...);
 Expr* new_Expr(const char* format, va_list vl);
 
-GenRet codegenValue(GenRet r);
-GenRet codegenValuePtr(GenRet r);
 #ifdef HAVE_LLVM
 llvm::Value* createTempVarLLVM(llvm::Type* type, const char* name);
 llvm::Value* createTempVarLLVM(llvm::Type* type);
 #endif
+
+GenRet codegenValue(GenRet r);
+GenRet codegenValuePtr(GenRet r);
+
 GenRet createTempVarWith(GenRet v);
 
 GenRet codegenDeref(GenRet toDeref);
